fix(dcc): diagnose same-scope redefs, fix multidim typedef sizeof, printf field padding
Compiler (src/dcc/dcc_func.c)
- Diagnose same-scope variable redefinition. scan_local_decl_after_type()
  now reports "redefinition of 'name'" when find_local_decl() matches in the
  innermost open block (a genuine same-scope duplicate per C89 6.1.2.2),
  instead of silently swallowing it and keeping the first declaration's type.
  Legal shadowing in nested blocks and for-init scopes is unaffected.
- Fix sizeof for multidimensional array typedefs. parse_typedef_decl() now
  folds every inner dimension into the total element count, so
  `typedef T A[2][3]` yields sizeof(A) == sizeof(T) * 6 rather than only the
  first dimension.

Runtime (DCCRTL.MAC)
- Correct printf integer field padding: add pf_iprepad/pf_ipostpad helpers so
  left-justified ('-') integer conversions emit trailing spaces and the field
  width is honored consistently.

Host validation harness (scripts/validate-unit-test.ps1)
- Add support for the new "requires-32bit-linux-host-compiler" override. When
  a 32-bit Linux compiler is detected (gcc -m32 makes long/pointer 4 bytes),
  tests flagged with this property are run instead of host-skipped, since
  -m32 reproduces dcc's 32-bit long width.

Unit tests
- Compiler fix fallout: remove now-rejected duplicate declarations that relied
  on the old quirk (tret.c, tlongaud.c); baselines unchanged.
- Portability fixes so tests pass host validation on a 64-bit host:
  * primes/tap/tarray/ttrig: add missing <stdlib.h>/<string.h> includes
    (dcc tolerated implicit declarations; gnu99 hosts do not).
  * tarray: cast int32_t/uint32_t to (unsigned)long for %ld/%lu.
  * tptrdiff/tstr: rename the custom `wchar_t` typedef to `twchar_t` to avoid
    clashing with the host built-in.
  * tstr/tstring: make rand()-driven display loops deterministic so host and
    CP/M output match; regenerate affected baselines.
  * tstruct: replace conflicting K&R extern decls with <string.h>; fix a void
    function returning a value.
  * treg/tstrconv: build long-width boundary values from limits macros / mask
    32-bit results so they hold on both 32- and 64-bit long.
- New 32-bit-portable equivalents of intrinsically 16-bit tests, comparing
  width-masked bit patterns: tpromo32, tbits32, ttype32, tunary32 (+baselines).
- tests/_test_overrides.json: host-skip tests whose mismatches are intrinsic
  dcc/CP/M ABI or runtime differences (16-bit int, no struct padding, 32-bit
  long wrap, float precision/no inf-nan, CP/M file/heap, strerror text,
  argv[0]); flag the pure-32-bit-long tests (tlimits, tlong, tlongaud,
  tlongreg, tmuldiv) as requires-32bit-linux-host-compiler.

Misc
- .gitignore: ignore trw.dat/trw2.dat artifacts created by the trw* file-I/O
  tests when run on the host.

+++ tests/tarray.c
@@ -3,6 +3,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <string.h>
 
 #define _countof( X ) ( sizeof( X ) / sizeof( X[0] ) )
 
@@ -180,8 +181,8 @@ int main( int argc, char * argv[] )
 
     for ( int i = 0; i < _countof( au8 ); i++ )
     {
-        printf( "index %u, au8 %u, au16 %u, au32 %lu\n", i, au8[ i ], au16[ i ], au32[ i ] );
-        printf( "index %u, a8 %d, a16 %d, a32 %ld\n", i, a8[ i ], a16[ i ], a32[ i ] );
+        printf( "index %u, au8 %u, au16 %u, au32 %lu\n", i, au8[ i ], au16[ i ], (unsigned long) au32[ i ] );
+        printf( "index %u, a8 %d, a16 %d, a32 %ld\n", i, a8[ i ], a16[ i ], (long) a32[ i ] );
     }
 
     for ( int i = 0; i < 8; i++ )
@@ -197,8 +198,8 @@ int main( int argc, char * argv[] )
 
     for ( int i = 0; i < 8; i++ )
     {
-        printf( "index %u, au8 %u, au16 %u, au32 %lu\n", i, au8[ i ], au16[ i ], au32[ i ] );
-        printf( "index %u, a8 %d, a16 %d, a32 %ld\n", i, a8[ i ], a16[ i ], a32[ i ] );
+        printf( "index %u, au8 %u, au16 %u, au32 %lu\n", i, au8[ i ], au16[ i ], (unsigned long) au32[ i ] );
+        printf( "index %u, a8 %d, a16 %d, a32 %ld\n", i, a8[ i ], a16[ i ], (long) a32[ i ] );
     }
 
     printf( "ac: %c %c %c %c %c %c %c %c\n", ac[ 0 ], ac[ 1 ], ac[ 2 ], ac[ 3 ], ac[ 4 ], ac[ 5 ], ac[ 6 ], ac[ 7 ] );
